fix: Address Sanath/Jacob feedback on tests folder/package.json
