+++ samcli/local/init/templates/cookiecutter-aws-sam-hello-nodejs/{{cookiecutter.project_name}}/hello_world/package.json
@@ -2,14 +2,12 @@
   "name": "hello_world",
   "version": "1.0.0",
   "description": "hello world sample for NodeJS",
-  "main": "src/index.js",
-  "repository": "https://github.com/awslabs/cookiecutter-aws-sam-hello-nodejs",
   "author": "SAM CLI",
   "license": "MIT",
   "scripts": {
     "test": "mocha tests/unit/",
-    "start": "sam local start-api",
-    "invoke": "sam local invoke HelloWorldFunction -e event.json"
+    "start": "cd .. && sam local start-api",
+    "invoke": "cd .. && sam local invoke HelloWorldFunction -e event.json"
   },
   "devDependencies": {
     "chai": "^4.1.2",
