--- samcli/local/init/templates/cookiecutter-aws-sam-hello-nodejs/{{cookiecutter.project_name}}/tests/unit/test_hello_world.js
@@ -1,43 +1,0 @@
-'use strict';
-
-const app = require('../../hello_world/app.js');
-const chai = require('chai');
-const expect = chai.expect;
-var event, context;
-
-{% if cookiecutter.runtime == 'nodejs6.10' or cookiecutter.runtime == 'nodejs4.3' %}
-describe('Tests Handler', function () {
-    it('verifies successful response', function (done) {
-        app.lambdaHandler(event, context, function (err, result) {
-            try {
-                expect(result).to.be.an('object');
-                expect(result.statusCode).to.equal(200);
-                expect(result.body).to.be.an('string');
-
-                let response = JSON.parse(result.body);
-
-                expect(response).to.be.an('object');
-                expect(response.message).to.be.equal("hello world");
-                done();
-            } catch (e) {
-                done(e);
-            }
-        });
-    });
-});
-{% else %}
-describe('Tests index', function () {
-    it('verifies successful response', async () => {
-        const result = await app.lambdaHandler(event, context)
-
-        expect(result).to.be.an('object');
-        expect(result.statusCode).to.equal(200);
-        expect(result.body).to.be.an('string');
-
-        let response = JSON.parse(result.body);
-
-        expect(response).to.be.an('object');
-        expect(response.message).to.be.equal("hello world");
-    });
-});
-{% endif %}
